+++ web-lgsm.py
@@ -30,7 +30,7 @@ def run_command_popen(command):
             command,
             shell=True,
             executable="/bin/bash",
-            stdin=subprocess.PIPE,
+            stdin=None,   # Connect input to the terminal.
             stdout=None,  # Direct output to the terminal.
             stderr=None,  # Direct error output to the terminal.
             text=True,
@@ -87,6 +87,7 @@ import getopt
 import shutil
 import string
 import getpass
+import tarfile
 import configparser
 from dotenv import load_dotenv
 from werkzeug.security import generate_password_hash
@@ -116,7 +117,7 @@ except KeyError as e:
 os.environ["LOG_LEVEL"] = LOG_LEVEL
 
 # Global options hash.
-O = {"verbose": False, "check": False, "auto": False, "test_full": False}
+O = {"verbose": False, "check": False, "auto": False, "test_full": False, "noback": False}
 
 def stop_server():
     result = subprocess.run(["pkill", "gunicorn"], capture_output=True)
@@ -439,7 +440,7 @@ def backup_file(filename):
     return backup_filename
 
 
-def backup_dir(dirname, tar=None):
+def backup_dir(dirname, tar=False):
     """Back's up directories using shutil.copydirtree, optionally tar's them too"""
     if not os.path.isdir(dirname):
         print(
@@ -449,89 +450,72 @@ def backup_dir(dirname, tar=None):
 
     epoc = int(time.time())
     backup_dirname = f"{dirname}.{epoc}.bak"
-    shutil.copytree(dirname, backup_dirname)
     print(f" [*] Backing up {dirname} to {backup_dirname}")
+    shutil.copytree(dirname, backup_dirname)
 
     if tar:
         tar_filename = f"{backup_dirname}.tar.gz"
+        print(f" [*] Creating tar file {tar_filename}")
         with tarfile.open(tar_filename, "w:gz") as tar_handle:
             tar_handle.add(backup_dirname, arcname=os.path.basename(backup_dirname))
-        print(f" [*] Creating tar file {tar_filename}")
         shutil.rmtree(backup_dirname)
         return tar_filename
 
     return backup_dirname
 
+def is_up_to_date():
+    """
+    Check's if web-lgsm already up-to-date or not.
+    """
+    run_command('git fetch --quiet')
+    output = run_command('git status -sb')
+    if 'ahead' in output or 'behind' in output:
+        return False
+
+    return True
+
 
 def update_weblgsm():
-    # Updates broken right now cause I suck a programming. Already have a todo
-    # to fix it. Marking this broken for the meantime.
-    print("Sorry, update is broken right now :( Will fix soon. In the meantime just git pull or clone the newest version and reinstall.")
-    return
-
-    local, remote, base = get_git_info()
-
-    if local == remote:
+    """
+    Update's the web-lgsm itself.
+    """
+    if is_up_to_date():
         print(" [*] Web LGSM already up to date!")
         return
 
-    elif local == base:
-        print(" [!] Update Required!")
-        if O["check"]:
-            return
-
-        if not O["auto"]:
-            resp = input(" [*] Would you like to update now? (y/n): ")
-            if resp.lower() != "y":
-                exit()
-
-        # Backup whole web-lgsm folder.
-        backup_dir(SCRIPTPATH, True)
-        backup_file("main.conf")
-
-        print(" [*] Pulling update from github...")
-        run_command("git clean -f")
-        run_command("git pull")
-
-        epoc = int(time.time())
-        print(f" [*] Backing up venv to venv.{epoc}.bak...")
-        backup_dir("venv")
-
-        print(" [*] Reinstalling newest web-lgsm...")
-        install_script = os.path.join(SCRIPTPATH, 'install.sh')
-        run_command(f"{install_script} -d")
-
-        print(" [*] Checking/Updating database fields...")
-        update_db = os.path.join(SCRIPTPATH, 'scripts/update-db.sh')
-        run_command(f"{update_db} -d")
-
-        print(" [*] Update Complete!")
+    print(" [!] Update Available!")
+    if O["check"]:
         return
 
-    elif remote == base:
-        print(" [!] Local ahead of remote, need push?", file=sys.stderr)
-        print(" [-] Note: Normal users should not see this.", file=sys.stderr)
-        sys.exit(1)
+    if not O["auto"]:
+        resp = input(" [*] Would you like to update now? (y/n): ")
+        if resp.lower() != "y":
+            exit()
 
-    print(" [!] Something has gone horribly wrong!", file=sys.stderr)
-    print(" [-] It's possible your local repo has diverged.", file=sys.stderr)
-    sys.exit(2)
+    # Backup whole web-lgsm folder.
+    if not O["noback"]:
+        print(" [*] Backing up and tarring, this may take a while...")
+        backup_dir(SCRIPTPATH, tar=True)
+        backup_file("main.conf")
 
+    print(" [*] Pulling update from github...")
+    run_command("git fetch --all")
+    run_command("git reset --hard origin/master")
 
-def check_sudo():
-    try:
-        # Run the sudo command with the -v option to validate the current timestamp.
-        result = subprocess.run(
-            ["sudo", "-v"], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-        if result.returncode != 0:
-            print(
-                " [*] No active sudo tty ticket. Please run the script with an active sudo session."
-            )
-            sys.exit(1)
-    except Exception as e:
-        print(f" [!] An error occurred while checking sudo status: {e}")
-        sys.exit(1)
+    print(" [*] Uninstalling old web-lgsm...")
+    install_script = os.path.join(SCRIPTPATH, 'uninstall.sh')
+    run_command(f"{uninstall_script} -d")
+
+    print(" [*] Reinstalling newest web-lgsm...")
+    install_script = os.path.join(SCRIPTPATH, 'install.sh')
+    run_command(f"{install_script} -d")
+
+    print(" [*] Updating database...")
+    run_command("flask --app app:main db upgrade")
+
+    # Green check!
+    print(f" [\033[32m✓\033[0m] Update Complete!")
+    return
 
 
 def run_tests():
@@ -559,12 +543,10 @@ def run_tests():
                 # Then torch dir.
                 shutil.rmtree(mcdir)
 
-#            cmd = "python -m pytest -v --maxfail=1"
             cmd = "coverage run -m pytest -v"
             run_command_popen(cmd)
 
         else:
-#            cmd = "python -m pytest -vvv -k 'not test_install_newuser and not test_install_sameuser' --maxfail=1"
             cmd = "coverage run -m pytest -vvv -k 'not test_install_newuser and not test_install_sameuser'"
             run_command_popen(cmd)
 
@@ -597,6 +579,7 @@ def print_help():
   ║   -p, --passwd        Change web user password           ║
   ║   -u, --update        Update web-lgsm version            ║
   ║   -c, --check         Check if an update is available    ║
+  ║   -n, --noback        Don't backup web-lgsm for updates  ║
   ║   -a, --auto          Run an auto update                 ║
   ║   -f, --fetch_json    Fetch latest game servers json     ║
   ║   -t, --test          Run project's pytest tests (short) ║
@@ -621,13 +604,14 @@ def main(argv):
             "passwd",
             "update",
             "check",
+            "noback",
             "auto",
             "fetch_json",
             "test",
             "test_full",
             "valid=",
         ]
-        opts, args = getopt.getopt(argv, "hsmrqdvpucaftxj:", longopts)
+        opts, args = getopt.getopt(argv, "hsmrqdvpucnaftxj:", longopts)
     except getopt.GetoptError as e:
         print(e)
         print_help()
@@ -647,6 +631,8 @@ def main(argv):
             O["auto"] = True
         if opt in ("-x", "--test_full"):
             O["test_full"] = True
+        if opt in ("-n", "--noback"):
+            O["noback"] = True
 
     # Do the needful based on opts.
     for opt, arg in opts: